+++ src/addresses.h
@@ -215,6 +215,10 @@
 #define RCT2_ADDRESS_NEW_WINDOW_PTR					0x014234B8
 #define RCT2_ADDRESS_VIEWPORT_LIST					0x014234BC
 #define RCT2_ADDRESS_NEW_VIEWPORT_PTR				0x01423570
+#define RCT2_ADDRESS_WINDOW_GUEST_LIST_SELECTED_FILTER		0x00F1EE06
+#define RCT2_ADDRESS_WINDOW_GUEST_LIST_SELECTED_TAB	0x00F1EE12
+#define RCT2_ADDRESS_WINDOW_GUEST_LIST_SELECTED_VIEW	0x00F1EE13
+
 
 #define RCT2_ADDRESS_OS_TIME_MINUTE					0x01424654
 #define RCT2_ADDRESS_OS_TIME_HOUR					0x01424656
