+++ src/audio.h
@@ -48,6 +48,8 @@ int sound_prepare(int sound_id, rct_sound *sound, int var_8, int var_c);
 void sound_play_panned(int sound_id, int x);
 int sound_play(rct_sound* sound, int looping, int volume, int pan, int frequency);
 void sound_stop(rct_sound *sound);
+void pause_sounds();
+void unpause_sounds();
 
 typedef enum {
 	RCT2_SOUND_SCREAM = 11,
