+++ src/editor.c
@@ -29,6 +29,9 @@
 #include "ride.h"
 #include "window.h"
 #include "viewport.h"
+#include "finance.h"
+#include "audio.h"
+#include "strings.h"
 
 static void set_all_land_owned();
 
@@ -49,12 +52,12 @@ void editor_load()
 	reset_saved_strings();
 	RCT2_CALLPROC_EBPSAFE(0x0069EB13); // sprites_init
 	ride_init_all();
-	RCT2_CALLPROC_EBPSAFE(0x0068F083); // window_guest_list_init_vars_a
+	window_guest_list_init_vars_a();
 	RCT2_CALLPROC_EBPSAFE(0x006BD3A4);
 	park_init();
 	finance_init();
 	date_reset();
-	RCT2_CALLPROC_EBPSAFE(0x0068F050); // window_guest_list_init_vars_b
+	window_guest_list_init_vars_b();
 	RCT2_CALLPROC_EBPSAFE(0x006BD39C);
 	RCT2_GLOBAL(RCT2_ADDRESS_SCREEN_FLAGS, uint8) = SCREEN_FLAGS_SCENARIO_EDITOR;
 	RCT2_GLOBAL(0x0141F570, uint8) = 0;
@@ -98,12 +101,12 @@ void trackdesigner_load()
 	reset_saved_strings();
 	RCT2_CALLPROC_EBPSAFE(0x0069EB13); // reset_sprites
 	ride_init_all();
-	RCT2_CALLPROC_EBPSAFE(0x0068F083); // window_guest_list_init_vars_a
+	window_guest_list_init_vars_a();
 	RCT2_CALLPROC_EBPSAFE(0x006BD3A4);
 	park_init();
 	finance_init();
 	date_reset();
-	RCT2_CALLPROC_EBPSAFE(0x0068F050); // window_guest_list_init_vars_b
+	window_guest_list_init_vars_b();
 	RCT2_CALLPROC_EBPSAFE(0x006BD39C);
 	RCT2_GLOBAL(RCT2_ADDRESS_SCREEN_FLAGS, uint8) = SCREEN_FLAGS_TRACK_DESIGNER;
 	RCT2_GLOBAL(0x0141F570, uint8) = 0;
@@ -136,12 +139,12 @@ void trackmanager_load()
 	reset_saved_strings();
 	RCT2_CALLPROC_EBPSAFE(0x0069EB13); // reset_sprites
 	ride_init_all();
-	RCT2_CALLPROC_EBPSAFE(0x0068F083); // window_guest_list_init_vars_a
+	window_guest_list_init_vars_a();
 	RCT2_CALLPROC_EBPSAFE(0x006BD3A4);
 	park_init();
 	finance_init();
 	date_reset();
-	RCT2_CALLPROC_EBPSAFE(0x0068F050); // window_guest_list_init_vars_b
+	window_guest_list_init_vars_b();
 	RCT2_CALLPROC_EBPSAFE(0x006BD39C);
 	RCT2_GLOBAL(RCT2_ADDRESS_SCREEN_FLAGS, uint8) = SCREEN_FLAGS_TRACK_MANAGER;
 	RCT2_GLOBAL(0x0141F570, uint8) = 0;
@@ -167,87 +170,4 @@ static void set_all_land_owned()
 	int mapSize = RCT2_GLOBAL(RCT2_ADDRESS_MAP_SIZE, sint16);
 
 	game_do_command(64, 1, 64, 2, 56, (mapSize - 2) * 32, (mapSize - 2) * 32);
-}
-
-/**
-*
-*  rct2: 0x00667104
-*/
-void reset_park_entrances() {
-	RCT2_GLOBAL(0x013573D4, uint16) = 0;
-
-	for (short i = 0; i < 4; i++) {
-		RCT2_ADDRESS(RCT2_ADDRESS_PARK_ENTRANCE_X, uint16)[i] = 0x8000;
-	}
-
-	RCT2_GLOBAL(0x013573F2, uint16) = 0xFFFF;
-	RCT2_GLOBAL(0x013573F8, uint16) = 0xFFFF;
-}
-
-/**
-*
-*  rct2: 0x006C4209
-*/
-void reset_saved_strings() {
-	for (int i = 0; i < 1024; i++) {
-		RCT2_ADDRESS(0x135A8F4, uint8)[i * 32] = 0;
-	}
-}
-
-/**
-*
-*  rct2: 0x006BABB4
-*/
-void pause_sounds() {
-	if (++RCT2_GLOBAL(0x009AF59C, uint8) == 1) {
-		RCT2_CALLPROC_EBPSAFE(0x006BCAE5);
-		RCT2_CALLPROC_EBPSAFE(0x006BABDF);
-		RCT2_CALLPROC_EBPSAFE(0x006BCA9F);
-		RCT2_CALLPROC_EBPSAFE(0x006BD07F);
-	}
-}
-
-/**
-*
-*  rct2: 0x006BABD8
-*/
-void unpause_sounds() {
-	RCT2_GLOBAL(0x009AF59C, uint8)--;
-}
-
-/**
-*
-*  rct2: 0x0069DEFB
-*/
-void finance_init() {
-
-	for (short i = 0; i < 56; i++) {
-		RCT2_ADDRESS(0x01357848, uint32)[i] = 0;
-	}
-
-	RCT2_GLOBAL(0x0135832C, uint32) = 0;
-
-	RCT2_GLOBAL(RCT2_ADDRESS_CURRENT_PROFIT, uint32) = 0;
-
-	RCT2_GLOBAL(0x01358334, uint32) = 0;
-	RCT2_GLOBAL(0x01358338, uint16) = 0;
-
-	RCT2_GLOBAL(0x013573DC, uint32) = 100000;
-
-	RCT2_GLOBAL(RCT2_ADDRESS_CURRENT_MONEY_ENCRYPTED, sint32) = ENCRYPT_MONEY(100000);
-	RCT2_GLOBAL(RCT2_ADDRESS_CURRENT_LOAN, sint32) = 100000;
-	RCT2_GLOBAL(RCT2_ADDRESS_MAXIMUM_LOAN, uint32) = 200000;
-
-	RCT2_GLOBAL(0x013587D0, uint32) = 0;
-
-	RCT2_GLOBAL(RCT2_ADDRESS_CURRENT_INTEREST_RATE, uint8) = 10;
-	RCT2_GLOBAL(RCT2_ADDRESS_CURRENT_PARK_VALUE, sint32) = 0;
-	RCT2_GLOBAL(RCT2_ADDRESS_CURRENT_COMPANY_VALUE, sint32) = 0;
-	RCT2_GLOBAL(RCT2_ADDRESS_COMPLETED_COMPANY_VALUE, sint32) = 0x80000000;
-	RCT2_GLOBAL(RCT2_ADDRESS_TOTAL_ADMISSIONS, uint32) = 0;
-	RCT2_GLOBAL(RCT2_ADDRESS_INCOME_FROM_ADMISSIONS, uint32) = 0;
-
-	RCT2_GLOBAL(0x013587D8, uint16) = 0x3F;
-
-	RCT2_CALLPROC_EBPSAFE(0x0069E869);
 }
