+++ src/finance.c
@@ -101,3 +101,40 @@ void finance_pay_interest()
 
 	finance_payment((sint32)tempcost, RCT_EXPENDITURE_TYPE_INTEREST);
 }
+
+/**
+*
+*  rct2: 0x0069DEFB
+*/
+void finance_init() {
+
+	for (short i = 0; i < 56; i++) {
+		RCT2_ADDRESS(0x01357848, uint32)[i] = 0;
+	}
+
+	RCT2_GLOBAL(0x0135832C, uint32) = 0;
+
+	RCT2_GLOBAL(RCT2_ADDRESS_CURRENT_PROFIT, uint32) = 0;
+
+	RCT2_GLOBAL(0x01358334, uint32) = 0;
+	RCT2_GLOBAL(0x01358338, uint16) = 0;
+
+	RCT2_GLOBAL(0x013573DC, uint32) = 100000;
+
+	RCT2_GLOBAL(RCT2_ADDRESS_CURRENT_MONEY_ENCRYPTED, sint32) = ENCRYPT_MONEY(100000);
+	RCT2_GLOBAL(RCT2_ADDRESS_CURRENT_LOAN, sint32) = 100000;
+	RCT2_GLOBAL(RCT2_ADDRESS_MAXIMUM_LOAN, uint32) = 200000;
+
+	RCT2_GLOBAL(0x013587D0, uint32) = 0;
+
+	RCT2_GLOBAL(RCT2_ADDRESS_CURRENT_INTEREST_RATE, uint8) = 10;
+	RCT2_GLOBAL(RCT2_ADDRESS_CURRENT_PARK_VALUE, sint32) = 0;
+	RCT2_GLOBAL(RCT2_ADDRESS_CURRENT_COMPANY_VALUE, sint32) = 0;
+	RCT2_GLOBAL(RCT2_ADDRESS_COMPLETED_COMPANY_VALUE, sint32) = 0x80000000;
+	RCT2_GLOBAL(RCT2_ADDRESS_TOTAL_ADMISSIONS, uint32) = 0;
+	RCT2_GLOBAL(RCT2_ADDRESS_INCOME_FROM_ADMISSIONS, uint32) = 0;
+
+	RCT2_GLOBAL(0x013587D8, uint16) = 0x3F;
+
+	RCT2_CALLPROC_EBPSAFE(0x0069E869);
+}
