+++ src/strings.h
@@ -25,6 +25,7 @@ typedef unsigned short rct_string_id;
 
 void format_string(char *dest, rct_string_id format, void *args);
 void generate_string_file();
+void reset_saved_strings();
 
 enum {
 	FORMAT_TINYFONT = 7,
